# docs/swagger.yaml
basePath: /
definitions:
  domain.Job:
    description: Job information for task processing
    properties:
      attempts:
        example: 0
        type: integer
      completed_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      group_id:
        example: group-123
        type: string
      id:
        example: 1
        type: integer
      idempotency_key:
        example: unique-key-123
        type: string
      lease_expires_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      lease_token:
        type: integer
      metadata:
        $ref: '#/definitions/domain.Metadata'
      payload:
        example: '{"task": "data"}'
        type: string
      queue_id:
        example: 1
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/domain.JobStatus'
        enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED
        example: PENDING
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  domain.JobStatus:
    enum:
    - PENDING
    - PROCESSING
    - COMPLETED
    - FAILED
    type: string
    x-enum-varnames:
    - JobStatusPending
    - JobStatusProcessing
    - JobStatusCompleted
    - JobStatusFailed
  domain.Metadata:
    properties:
      consumer_id:
        example: consumer-123
        type: string
      last_pull_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  domain.Queue:
    description: Queue configuration for task processing
    properties:
      created_at:
        description: |-
          DLQInterval   time.Duration `json:"dlq_interval"`
          DLQId         int64         `json:"dlq_id"`
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      lease_duration:
        description: Duration in Go format (e.g., "30s", "1m")
        example: 30s
        type: string
      max_attempts:
        example: 3
        type: integer
      name:
        example: default-queue
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  rest.CompleteJobRequest:
    properties:
      lease_token:
        example: 1
        type: integer
      queue_id:
        example: 1
        type: integer
    type: object
  rest.ProcessJobRequest:
    properties:
      queue_id:
        example: 1
        type: integer
    type: object
host: localhost:9090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a distributed task queue API server built with Go and Echo
    framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Distributed Task Queue API
  version: "1.0"
paths:
  /jobs:
    get:
      consumes:
      - application/json
      description: Retrieve pending jobs from the queue
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Job'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Pull pending jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Create and publish a new job to the queue
      parameters:
      - description: Job object
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/domain.Job'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Job'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Publish a new job
      tags:
      - jobs
  /jobs/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark a job as completed using its ID, lock token, and queue ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Complete Job Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.CompleteJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Job'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Complete a job
      tags:
      - jobs
  /jobs/process:
    post:
      consumes:
      - application/json
      description: Process and lock a job from the specified queue
      parameters:
      - description: Process Job Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.ProcessJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Job'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Process a job from queue
      tags:
      - jobs
  /queues:
    get:
      consumes:
      - application/json
      description: Retrieve all queues
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Queue'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all queues
      tags:
      - queues
    post:
      consumes:
      - application/json
      description: Create a new queue with specified configuration
      parameters:
      - description: Queue object
        in: body
        name: queue
        required: true
        schema:
          $ref: '#/definitions/domain.Queue'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Queue'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new queue
      tags:
      - queues
  /queues/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a queue by its ID
      parameters:
      - description: Queue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Queue'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get queue by ID
      tags:
      - queues
schemes:
- http
- https
swagger: "2.0"